nanoclaw/requirements.txt
# ============================================================
# NanoClaw — Python Dependencies (Fase 7)
# ============================================================
# Kjernebiblioteker for AI-agenten som kjører i hver
# brukercontainer.
# ============================================================

# OpenAI SDK — kompatibel med LiteLLM proxy
openai>=1.30.0

# Google API-klienter — Gmail og Calendar (read-only)
google-api-python-client>=2.130.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0

# HTTP-klient for Vault/Orkestrator-kommunikasjon
httpx>=0.27.0

# Watchdog — filsystem-overvåking for wake-signal
watchdog>=4.0.0
